Move PagingState[Response] to scylla-cql-core
