+++ scylla-cql-core/src/frame/request/mod.rs
@@ -1,0 +1,1 @@
+pub mod query;
